Fixed Movement Issue
Movement for other Players should work now fine

+++ src/shared/StackBuffer.h
@@ -31,7 +31,7 @@ class SERVER_DECL StackBufferBase
     public:
         virtual ~StackBufferBase() {}
         virtual uint8* GetBufferPointer() = 0;
-        virtual uint16 GetOpcode() { return 0; }
+        virtual uint32 GetOpcode() { return 0; }
         virtual uint32 GetSize() = 0;
 };
